Add version endpoint.
