+++ config/config.conf
@@ -6,6 +6,7 @@ httpHost="0.0.0.0"
 httpPort=7066
 httpHealthCheckPath="/ping"
 httpStatusPath="/status"
+httpVersionPath="/version"
 aggregationHost="0.0.0.0"
 aggregationPort=7065
 jvmMetricsCollectionInterval="PT.5S"
