+++ src/main/resources/version.json
